Organize storage handlers
