+++ pkg/storage/mock/mock.go
@@ -18,8 +18,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/coreos-inc/quartermaster/pkg/operator"
 	"github.com/coreos-inc/quartermaster/pkg/spec"
+	qmstorage "github.com/coreos-inc/quartermaster/pkg/storage"
+
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/levels"
 
@@ -38,13 +39,13 @@ func init() {
 		With("ts", log.DefaultTimestampUTC, "caller", log.DefaultCaller))
 }
 
-func New(client clientset.Interface, qm restclient.Interface) (operator.StorageType, error) {
+func New(client clientset.Interface, qm restclient.Interface) (qmstorage.StorageType, error) {
 	s := &MockStorage{
 		client: client,
 		qm:     qm,
 	}
 
-	return &operator.StorageHandlerFuncs{
+	return &qmstorage.StorageHandlerFuncs{
 		StorageHandler:     s,
 		TypeFunc:           s.Type,
 		InitFunc:           s.Init,
